requirements.txt
Django==1.7.1
resolver==0.2.1
selector==0.8.11
Paste==1.7.2
PasteDeploy==1.3.3
PasteScript==1.7.3
httplib2==0.7.4
feedparser==4.1
Markdown==2.3.1
uuid==1.30
psycopg2==2.5.1
nose==1.3.0
versiontools==1.8.2
statsd==2.0.2
flake8==1.3.1
pyparsing==2.0.1
tornado==3.1
BeautifulSoup==3.2.1
cssselect==0.8
lxml==3.2.3
fuzzywuzzy==0.2
sure==1.2.2
ipython==0.13.2
ipdb==0.7
rdflib==3.1.0
selenium==2.31.0
logilab-common==0.59.0
logilab-astng==0.24.1
pylint==0.26.0
pep8==1.4
pyflakes==0.5.0
coverage==3.5.3
pyzmq==13.1.0
pytz==2013b
six==1.2.0

djangowind==0.13.2
django-indexer==0.2.1
django-templatetag-sugar==0.1
django-paging==0.2.2
django-annoying==0.8.0
django-nose==1.2
django-appconf==0.6
django_compressor==1.4
django-statsd-mozilla==0.3.11
raven==4.0.3
django-bootstrap-form==2.0.3
django-jenkins==0.16.3
django-waffle==0.9.0
django-smoketest==0.3.0
django-markwhat==1.0
